proompter-master/requirements.txt
opik>=1.7.36
litellm>=1.0.0
python-dotenv>=1.0.0 